resume_tests():  On Windows, this could try to do
    args.insert(1, '-O')

after `args` had been changed from a list to a string.
Blew up then.  Swapped the order of the code so that
list operations are all done while `args` is still a
list.

The bad news:  the tests run on Windows under Python
2.3.5 using -O then, but _sometimes_ Python itself dies
with a segfault.  No clues yet.

Have not seen that happen under 2.4.2 with -O.