Check that first word of exception message is a dotted name
As far as I can see, there's no way to do a perfect job of this because
of the way the doctest interface works: check_output just doesn't really
know when it's looking at a traceback.  However, this makes
renormalization of exceptions without arguments less disruptive to
projects that apply IGNORE_EXCEPTION_MODULE_IN_PYTHON2 to an entire
file, by checking whether the part of the exception message that looks
like the exception type is in fact a valid Python dotted name.

For good measure, also make IGNORE_EXCEPTION_MODULE_IN_PYTHON2 work more
like the standard doctest option flag IGNORE_EXCEPTION_DETAIL, by
checking both the stripped and unstripped versions of the example on
Python 2 and passing if either one matches.  This makes it less likely
for renormalization of exceptions without arguments to cause problems.